+++ backend/ddang/src/main/java/com/ddang/ddang/auction/domain/exception/InvalidPriceValueException.java
@@ -1,4 +1,4 @@
-package com.ddang.ddang.actuion.domain.exception;
+package com.ddang.ddang.auction.domain.exception;
 
 public class InvalidPriceValueException extends IllegalArgumentException {
