+++ backend/ddang/src/main/java/com/ddang/ddang/auction/infrastructure/persistence/JpaAuctionRepository.java
@@ -1,6 +1,6 @@
-package com.ddang.ddang.actuion.infrastructure.persistence;
+package com.ddang.ddang.auction.infrastructure.persistence;
 
-import com.ddang.ddang.actuion.domain.Auction;
+import com.ddang.ddang.auction.domain.Auction;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JpaAuctionRepository extends JpaRepository<Auction, Long> {
