+++ backend/ddang/src/test/java/com/ddang/ddang/auction/domain/PriceTest.java
@@ -1,6 +1,6 @@
-package com.ddang.ddang.actuion.domain;
+package com.ddang.ddang.auction.domain;
 
-import com.ddang.ddang.actuion.domain.exception.InvalidPriceValueException;
+import com.ddang.ddang.auction.domain.exception.InvalidPriceValueException;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
