+++ backend/ddang/src/test/java/com/ddang/ddang/auction/infrastructure/persistence/JpaAuctionRepositoryTest.java
@@ -1,7 +1,7 @@
-package com.ddang.ddang.actuion.infrastructure.persistence;
+package com.ddang.ddang.auction.infrastructure.persistence;
 
-import com.ddang.ddang.actuion.domain.Auction;
-import com.ddang.ddang.actuion.domain.Price;
+import com.ddang.ddang.auction.domain.Auction;
+import com.ddang.ddang.auction.domain.Price;
 import com.ddang.ddang.configuration.JpaConfiguration;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
